Collectible to player.cs
Collectible moved to player.cs.
