+++ Assets/Scripts/Entity/Player.cs
@@ -1,11 +1,21 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
+using UnityEngine.UI;
 
 namespace Assets.Scripts.Entity
 {
     class Player : AbstractEntity
     {
+
+        public int collectible;
+        public Text counter;
+
+        new public void Start() {
+            base.Start();
+            changeCounter();
+
+        }
         public void Update()
         {
             Movement = new Vector2(Input.GetAxis("Horizontal"), 0);
@@ -20,6 +30,10 @@ namespace Assets.Scripts.Entity
                      collectible = collectible + 1;
                      changeCounter();
                 }
+        }
+
+        public void changeCounter() {
+            counter.text = "Collectibles: " + collectible;
         }
     }
 }
