+++ ProjectSettings/TagManager.asset
@@ -5,6 +5,7 @@ TagManager:
   serializedVersion: 2
   tags:
   - Forground
+  - Collectible
   layers:
   - Default
   - TransparentFX
@@ -50,4 +51,7 @@ TagManager:
     locked: 0
   - name: Forground
     uniqueID: 2788644723
+    locked: 0
+  - name: Pickups
+    uniqueID: 549795201
     locked: 0
